Update marketplaceApps.json
I added static IP to both the Timpi Collector and GeoCore due to complains from their community about constant IP changes. This should make it better a bit. 

I also adjusted the prices because of it, and additionally lowered the storage requirement for the Collector Node from 3GB to 1GB as discussed with the timpi team.